Server readme update (#534)
Added documentation on how to start postgres and pgadmin for development. I initially tried the approach of installing postgres locally, but figured it is easier to run it in docker with pgadmin. Then the backend development server is started locally with node to start developing with hot reload.
Also modified the servers docker-compose.yaml to expose database to localhost:5432.

---------

Co-authored-by: Kuru Kimi <kuruk1@l23-0094.org.aalto.fi>